Move tenants to a separate module, move rsa key related methods to auth_controller
